corriguiendo errores de modularización en contact
