--- src/app/components/contact/contact.component.html
@@ -1,1 +1,0 @@
-<p>Contact</p>
