+++ src/app/contacts/contacts-routing.module.ts
@@ -4,7 +4,7 @@ import { ContactComponent } from './components/contact/contact.component';
 
 
 const routes: Routes = [
-    { path: 'contactos', component: ContactComponent},
+    { path: '', component: ContactComponent},
 ];
 
 @NgModule({
