+++ src/app/pages/components/server-internal/server-internal.component.ts
